[DevicesTableViewController] Renamed to DevicesViewController and integrates Overlay
