+++ PiRemote/PiRemote/OverlayManager.swift
@@ -16,10 +16,12 @@ extension CGRect{
 
 
 class OverlayManager: NSObject {
+
+
+
     
-    
-   static func createErrorOverlay(message: String) -> UIAlertController{
-    return self.createActionOverlay(title:"Oh no!", withMessage: message)
+    static func createErrorOverlay(message: String) -> UIAlertController{
+        return self.createActionOverlay(title:"Oh no!", withMessage: message)
     }
     
     static func createSucessOverlay(message: String) -> UIAlertController{
@@ -27,8 +29,8 @@ class OverlayManager: NSObject {
     }
     
     
-    static func createLoadingSpinner() -> UIAlertController{
-        let alert = UIAlertController(title: nil, message: "Please wait...", preferredStyle: .alert)
+    static func createLoadingSpinner(withMessage actionMessage: String = "Please wait...") -> UIAlertController{
+        let alert = UIAlertController(title: nil, message: actionMessage, preferredStyle: .alert)
         
         alert.view.tintColor = UIColor.black
         let loadingIndicator: UIActivityIndicatorView = UIActivityIndicatorView(frame: CGRect(10, 5, 50, 50)) as UIActivityIndicatorView
@@ -55,5 +57,4 @@ class OverlayManager: NSObject {
         return alertController
     }
     
-    }
-
+}
